working solution but super ugly
for too small timeouts, the patch routines is too slow and the wait function timeouts. So I added a minimal timeout at 0.5s value in the sweep_acquire function. It is a bit high and ideally you would want this timeout to be used only when the function is called behind a Qt slot. I'm just committing this solution as a backup because it has the great advantage of actually working !