Update async_utils.py
Remove prints and now it works with a shell, IPython and in jupyter notebook. Using asyncio.events._set_running_loop(LOOP) or asyncio.set_event_loop(LOOP) we can probably avoid patching the asyncio.wait and asyncio.sleep function but this seems to work for now.

TO DO : Test on a real RedPitaya and write a test function that tests the calibrate function both through a direct call and via a signal.